+++ src/Sulu/Bundle/SearchBundle/CHANGELOG.md
@@ -3,6 +3,6 @@ CHANGELOG
 
 dev-develop
 -----------
-
+- Deindex unpublished nodes
 - Map Url, Description and Title fields from structure template to index metadata
 - Event listener for setting the portal prefix on the document URL